requirements.txt
mlflow==2.14.3
joblib==1.4.2
scikit-learn==1.5.1
numpy ==1.26.4
PyYAML==6.0.1
pandas==2.2.2
imbalanced-learn==0.12.3
